// TitanFitness.Frontend/src/app/core/models/memberships/membership-details.model.ts
export interface MembershipDetailsRecord {
  membershipId: number;
  memberId: number;
  planId: number;
  planName: string;
  purchaseDate: string;
  startDate: string;
  endDate: string;
  status: string;
  pricePaid: number;
  durationInMonths: number;
  maximumFreezeDays: number;
  maximumNumberOfFreezes: number;
  guestPassQuota: number;
  accessScope: string;
}